# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CCOS - Credit Card Operating System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Card {
  id: number;
  name: string;
  bank?: string | null;
  limit: number;
  /**
   * @minimum 1
   * @maximum 31
   */
  statementDay: number;
  /**
   * @minimum 1
   * @maximum 31
   */
  dueDay: number;
  color?: string | null;
  used: number;
  utilization: number;
  safeSpendRemaining: number;
  daysUntilDue: number;
  createdAt: string;
}

export interface CreateCardBody {
  name: string;
  bank?: string | null;
  limit: number;
  /**
   * @minimum 1
   * @maximum 31
   */
  statementDay: number;
  /**
   * @minimum 1
   * @maximum 31
   */
  dueDay: number;
  color?: string | null;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  expense: "expense",
  payment: "payment",
  income: "income",
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  amount: number;
  date: string;
  cardId?: number | null;
  note?: string | null;
  cardName?: string | null;
  createdAt: string;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  expense: "expense",
  payment: "payment",
  income: "income",
} as const;

export interface CreateTransactionBody {
  type: CreateTransactionBodyType;
  amount: number;
  date: string;
  cardId?: number | null;
  note?: string | null;
}

export interface CardSummary {
  id: number;
  name: string;
  bank?: string | null;
  color?: string | null;
  limit: number;
  used: number;
  utilization: number;
  safeSpendRemaining: number;
  daysUntilDue: number;
  score: number;
}

export type WarningType = (typeof WarningType)[keyof typeof WarningType];

export const WarningType = {
  utilization: "utilization",
  due_soon: "due_soon",
  cash_buffer: "cash_buffer",
} as const;

export type WarningSeverity =
  (typeof WarningSeverity)[keyof typeof WarningSeverity];

export const WarningSeverity = {
  info: "info",
  warning: "warning",
  critical: "critical",
} as const;

export interface Warning {
  type: WarningType;
  message: string;
  cardId?: number | null;
  severity: WarningSeverity;
}

export interface DashboardSummary {
  bestCard?: CardSummary | null;
  cards: CardSummary[];
  warnings: Warning[];
  totalSafeSpendRemaining: number;
  totalUsed: number;
  totalLimit: number;
  cashBalance: number;
  recentTransactions: Transaction[];
}

export type ListTransactionsParams = {
  cardId?: number;
  type?: ListTransactionsType;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  expense: "expense",
  payment: "payment",
  income: "income",
} as const;