# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CCOS - Credit Card Operating System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all cards
 */
export const listCardsResponseStatementDayMax = 31;

export const listCardsResponseDueDayMax = 31;

export const ListCardsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  bank: zod.string().nullish(),
  limit: zod.number(),
  statementDay: zod.number().min(1).max(listCardsResponseStatementDayMax),
  dueDay: zod.number().min(1).max(listCardsResponseDueDayMax),
  color: zod.string().nullish(),
  used: zod.number(),
  utilization: zod.number(),
  safeSpendRemaining: zod.number(),
  daysUntilDue: zod.number(),
  createdAt: zod.string(),
});
export const ListCardsResponse = zod.array(ListCardsResponseItem);

/**
 * @summary Create a card
 */
export const createCardBodyStatementDayMax = 31;

export const createCardBodyDueDayMax = 31;

export const CreateCardBody = zod.object({
  name: zod.string(),
  bank: zod.string().nullish(),
  limit: zod.number(),
  statementDay: zod.number().min(1).max(createCardBodyStatementDayMax),
  dueDay: zod.number().min(1).max(createCardBodyDueDayMax),
  color: zod.string().nullish(),
});

/**
 * @summary Get a card by ID
 */
export const GetCardParams = zod.object({
  id: zod.coerce.number(),
});

export const getCardResponseStatementDayMax = 31;

export const getCardResponseDueDayMax = 31;

export const GetCardResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  bank: zod.string().nullish(),
  limit: zod.number(),
  statementDay: zod.number().min(1).max(getCardResponseStatementDayMax),
  dueDay: zod.number().min(1).max(getCardResponseDueDayMax),
  color: zod.string().nullish(),
  used: zod.number(),
  utilization: zod.number(),
  safeSpendRemaining: zod.number(),
  daysUntilDue: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Update a card
 */
export const UpdateCardParams = zod.object({
  id: zod.coerce.number(),
});

export const updateCardBodyStatementDayMax = 31;

export const updateCardBodyDueDayMax = 31;

export const UpdateCardBody = zod.object({
  name: zod.string(),
  bank: zod.string().nullish(),
  limit: zod.number(),
  statementDay: zod.number().min(1).max(updateCardBodyStatementDayMax),
  dueDay: zod.number().min(1).max(updateCardBodyDueDayMax),
  color: zod.string().nullish(),
});

export const updateCardResponseStatementDayMax = 31;

export const updateCardResponseDueDayMax = 31;

export const UpdateCardResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  bank: zod.string().nullish(),
  limit: zod.number(),
  statementDay: zod.number().min(1).max(updateCardResponseStatementDayMax),
  dueDay: zod.number().min(1).max(updateCardResponseDueDayMax),
  color: zod.string().nullish(),
  used: zod.number(),
  utilization: zod.number(),
  safeSpendRemaining: zod.number(),
  daysUntilDue: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a card
 */
export const DeleteCardParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List transactions
 */
export const ListTransactionsQueryParams = zod.object({
  cardId: zod.coerce.number().optional(),
  type: zod.enum(["expense", "payment", "income"]).optional(),
});

export const ListTransactionsResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum(["expense", "payment", "income"]),
  amount: zod.number(),
  date: zod.string(),
  cardId: zod.number().nullish(),
  note: zod.string().nullish(),
  cardName: zod.string().nullish(),
  createdAt: zod.string(),
});
export const ListTransactionsResponse = zod.array(ListTransactionsResponseItem);

/**
 * @summary Create a transaction
 */
export const CreateTransactionBody = zod.object({
  type: zod.enum(["expense", "payment", "income"]),
  amount: zod.number(),
  date: zod.string(),
  cardId: zod.number().nullish(),
  note: zod.string().nullish(),
});

/**
 * @summary Delete a transaction
 */
export const DeleteTransactionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get dashboard summary
 */
export const GetDashboardResponse = zod.object({
  bestCard: zod
    .object({
      id: zod.number(),
      name: zod.string(),
      bank: zod.string().nullish(),
      color: zod.string().nullish(),
      limit: zod.number(),
      used: zod.number(),
      utilization: zod.number(),
      safeSpendRemaining: zod.number(),
      daysUntilDue: zod.number(),
      score: zod.number(),
    })
    .nullish(),
  cards: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      bank: zod.string().nullish(),
      color: zod.string().nullish(),
      limit: zod.number(),
      used: zod.number(),
      utilization: zod.number(),
      safeSpendRemaining: zod.number(),
      daysUntilDue: zod.number(),
      score: zod.number(),
    }),
  ),
  warnings: zod.array(
    zod.object({
      type: zod.enum(["utilization", "due_soon", "cash_buffer"]),
      message: zod.string(),
      cardId: zod.number().nullish(),
      severity: zod.enum(["info", "warning", "critical"]),
    }),
  ),
  totalSafeSpendRemaining: zod.number(),
  totalUsed: zod.number(),
  totalLimit: zod.number(),
  cashBalance: zod.number(),
  recentTransactions: zod.array(
    zod.object({
      id: zod.number(),
      type: zod.enum(["expense", "payment", "income"]),
      amount: zod.number(),
      date: zod.string(),
      cardId: zod.number().nullish(),
      note: zod.string().nullish(),
      cardName: zod.string().nullish(),
      createdAt: zod.string(),
    }),
  ),
});