# lib/api-zod/src/generated/types/card.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CCOS - Credit Card Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface Card {
  id: number;
  name: string;
  bank?: string | null;
  limit: number;
  /**
   * @minimum 1
   * @maximum 31
   */
  statementDay: number;
  /**
   * @minimum 1
   * @maximum 31
   */
  dueDay: number;
  color?: string | null;
  used: number;
  utilization: number;
  safeSpendRemaining: number;
  daysUntilDue: number;
  createdAt: string;
}