# lib/api-zod/src/generated/types/cardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CCOS - Credit Card Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface CardSummary {
  id: number;
  name: string;
  bank?: string | null;
  color?: string | null;
  limit: number;
  used: number;
  utilization: number;
  safeSpendRemaining: number;
  daysUntilDue: number;
  score: number;
}