# lib/api-zod/src/generated/types/createTransactionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CCOS - Credit Card Operating System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransactionBodyType } from "./createTransactionBodyType";

export interface CreateTransactionBody {
  type: CreateTransactionBodyType;
  amount: number;
  date: string;
  cardId?: number | null;
  note?: string | null;
}