# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CCOS - Credit Card Operating System API
 * OpenAPI spec version: 0.1.0
 */
import type { CardSummary } from "./cardSummary";
import type { Transaction } from "./transaction";
import type { Warning } from "./warning";

export interface DashboardSummary {
  bestCard?: CardSummary | null;
  cards: CardSummary[];
  warnings: Warning[];
  totalSafeSpendRemaining: number;
  totalUsed: number;
  totalLimit: number;
  cashBalance: number;
  recentTransactions: Transaction[];
}