# lib/api-zod/src/generated/types/warning.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CCOS - Credit Card Operating System API
 * OpenAPI spec version: 0.1.0
 */
import type { WarningSeverity } from "./warningSeverity";
import type { WarningType } from "./warningType";

export interface Warning {
  type: WarningType;
  message: string;
  cardId?: number | null;
  severity: WarningSeverity;
}